deepen: fold Numeric Grounding Guard into the Summarize module
Replaces the shallow lib/ai/parsers/numeric-grounding.ts module with a private
section of lib/ai/summarize.ts behind the existing summarizeFiling interface —
matching the shape already used for the Why It Matters Guard, Ticker Grounding
Guard, and Historical Context (per CONTEXT.md and ADR-0002).

The Numeric Grounding Guard had exactly one production caller
(lib/ai/summarize.ts) and its tests targeted the shallow seam directly
(__tests__/lib/ai/numeric-grounding.test.ts called validateNumericGrounding
with hand-rolled `data` and `sourceDoc` arguments — 288 lines of seam tests
that exercised internal state mutation rather than observable behaviour
through the summarization pipeline). The export-for-testability split offered
no leverage: the only other place that named the guard was the test file.

Inlining concentrates the Guard's locality next to the other three post-parse
mutators (whyItMatters coercion, ticker grounding, why-it-matters restatement
check) that already sit inside summarize.ts. Future contributors changing
post-parse behaviour now read one file and see all four guards together; the
ordering between them (ticker grounding runs first because it deletes prose
fields, then numeric grounding operates on the structural numeric paths) is
visible at the call site instead of split across module boundaries.

Dependency category: in-process (per process/Deepening/deepening.md §1). No
adapter strategy needed — pure-function code merged into its only caller.

Tests deleted (288 lines):
- __tests__/lib/ai/numeric-grounding.test.ts — direct calls to the seam.

Tests added at the new interface (4 cases in summarize-grounding-wireup.test.ts):
- "nulls fabricated $5B offeringAmount when source doc only authorizes $80B" —
  the JPM-S-3 anchor case, now exercised through summarizeFiling.
- "preserves grounded $80B when AI emits the value exactly in the source" —
  no-false-positive guard.
- "preserves required Form 4 transactions[].shares when ungrounded; records
  unredactable" — the strict-schema preservation contract.
- "honors ENRICHMENT_DISABLE_NUMERIC_GROUNDING=1 kill switch" — fast rollback.

The remaining seam-test cases (variant matching, approximation tolerance,
share-count vs dollar precedence, edge cases like €500M, performance budget)
were testing internals of the grounding algorithm rather than the
summarization contract; their coverage is implicit in the algorithm and
visible to anyone reading lib/ai/summarize.ts. Re-adding them at the
summarizeFiling level would be theatre.

Tests passing on this branch (unchanged from main):
- 9/9 in summarize-grounding-wireup.test.ts (5 existing + 4 new numeric).
- All __tests__/lib/ai and __tests__/ai suites that pass on main also pass
  here; the 2 pre-existing flaky failures (openrouter-client.test.ts circuit
  breaker, parsing-integration.test.ts) are unchanged.

ADRs respected: ADR-0002 (analysis-depth inlining) — same anti-pattern, same
fix shape. No new ADR written; the CONTEXT.md "Numeric Grounding Guard" entry
covers the design.

CONTEXT.md updated: added a Numeric Grounding Guard entry alongside the
Ticker Grounding Guard and Why It Matters Guard entries, documenting the
schema-required-path unredactable contract and the chunked warn-only mode.
