--- __tests__/lib/ai/numeric-grounding.test.ts
@@ -1,288 +1,0 @@
-/**
- * Tests for the numeric grounding validator (PR 2).
- *
- * Anchor case: JPM S-3 fabricated `$5B` total when the actual prospectus
- * authorizes `$80B aggregate`. Validator should null the fabricated values
- * while preserving values that genuinely appear in the source doc.
- */
-
-import {
-  validateNumericGrounding,
-} from '../../../lib/ai/parsers/numeric-grounding';
-
-const PADDING = '\n' + 'context filler. '.repeat(80) + '\n';
-
-function pad(doc: string): string {
-  // Validator skips when sourceDoc < 1000 chars; pad with neutral filler.
-  return PADDING + doc + PADDING;
-}
-
-describe('validateNumericGrounding — passes through grounded values', () => {
-  it.each([
-    ['$5B exact match', '$5B', 'authorized up to $5B in securities'],
-    ['$5B vs $5,000,000,000 long form', '$5B', 'up to $5,000,000,000 aggregate'],
-    ['$5B vs $5 billion', '$5B', 'up to $5 billion aggregate'],
-    ['+34% vs "34%"', '+34%', 'rose 34% year over year'],
-    ['+34% vs "34 percent"', '+34%', 'rose 34 percent year over year'],
-    ['223,986 share count exact', '223,986 shares', 'sold 223,986 shares'],
-    ['1.2M shares vs 1,234,567 long form', '1.2M shares', 'sold 1,234,567 shares of common stock'],
-  ])('%s', (_label, value, doc) => {
-    const data: Record<string, unknown> = { offeringAmount: value };
-    const result = validateNumericGrounding(data, pad(doc), { formType: 'S-3' });
-    expect(result.rejectedFields).toEqual([]);
-    expect(data.offeringAmount).toBe(value);
-  });
-});
-
-describe('validateNumericGrounding — rejects ungrounded values', () => {
-  it.each([
-    ['$5B not in $80B doc', '$5B', 'authorized $80,000,000,000 aggregate'],
-    ['+34% not in doc', '+34%', 'no growth percentage stated in this filing'],
-  ])('%s', (_label, value, doc) => {
-    const data: Record<string, unknown> = { offeringAmount: value };
-    const result = validateNumericGrounding(data, pad(doc), { formType: 'S-3' });
-    expect(result.rejectedFields.length).toBe(1);
-    expect(data.offeringAmount).toBeNull();
-  });
-});
-
-describe('validateNumericGrounding — JPM regression', () => {
-  it('nulls fabricated $5B totalAuthorized when source has only $80B', () => {
-    const data: Record<string, unknown> = {
-      offeringAmount: '$5B',
-      shelfRegistration: {
-        totalAuthorized: '$5B',
-        remainingCapacity: '$5B',
-      },
-    };
-    const sourceDoc = pad(
-      'Up to $80,000,000,000, or the equivalent thereof in any other currency, of these securities ' +
-      'may be offered from time to time'
-    );
-    const result = validateNumericGrounding(data, sourceDoc, { formType: 'S-3' });
-    expect(result.rejectedFields.map((r) => r.path).sort()).toEqual([
-      'offeringAmount',
-      'shelfRegistration.remainingCapacity',
-      'shelfRegistration.totalAuthorized',
-    ]);
-    expect(data.offeringAmount).toBeNull();
-    expect((data.shelfRegistration as Record<string, unknown>).totalAuthorized).toBeNull();
-    expect((data.shelfRegistration as Record<string, unknown>).remainingCapacity).toBeNull();
-  });
-
-  it('preserves grounded $80B when AI emits the correct value', () => {
-    const data: Record<string, unknown> = {
-      shelfRegistration: { totalAuthorized: '$80B' },
-    };
-    const sourceDoc = pad('Up to $80,000,000,000 of these securities may be offered');
-    const result = validateNumericGrounding(data, sourceDoc, { formType: 'S-3' });
-    expect(result.rejectedFields).toEqual([]);
-    expect((data.shelfRegistration as Record<string, unknown>).totalAuthorized).toBe('$80B');
-  });
-});
-
-describe('validateNumericGrounding — array walking', () => {
-  it('nulls a single ungrounded financialHighlights[].change without dropping the row', () => {
-    // Note: 10-Q schema marks `value` as required (items.required), so an
-    // ungrounded value lands in unredactablePaths instead of being nulled.
-    // `change` is optional and IS nulled. Both behaviors are validated here.
-    const data: Record<string, unknown> = {
-      financialHighlights: [
-        { label: 'Revenue', value: '$50.5B', change: '+15%' }, // grounded
-        { label: 'Margin',  value: '$99B',   change: '+200%' }, // fabricated
-      ],
-    };
-    const sourceDoc = pad('Revenue of $50.5B grew 15%. Margin commentary in MD&A.');
-    const result = validateNumericGrounding(data, sourceDoc, { formType: '10-Q' });
-    const arr = data.financialHighlights as Array<Record<string, unknown>>;
-    expect(arr).toHaveLength(2);
-    expect(arr[1].change).toBeNull();
-    expect(arr[1].label).toBe('Margin');
-    expect(arr[0].value).toBe('$50.5B');
-    // `value` is required → preserved + reported in unredactablePaths.
-    expect(arr[1].value).toBe('$99B');
-    expect(result.unredactablePaths).toContain('financialHighlights[1].value');
-  });
-
-  it('nulls ungrounded value when the schema does not mark it required', () => {
-    // Use S-3 securityTypes[].amount which is not required.
-    const data: Record<string, unknown> = {
-      securityTypes: [
-        { type: 'Common Stock', amount: '$2B' },     // fabricated
-        { type: 'Debt Securities', amount: '$3B' },  // fabricated
-      ],
-    };
-    const sourceDoc = pad('authorized up to $80,000,000,000 aggregate, in any combination');
-    const result = validateNumericGrounding(data, sourceDoc, { formType: 'S-3' });
-    const arr = data.securityTypes as Array<Record<string, unknown>>;
-    expect(arr[0].amount).toBeNull();
-    expect(arr[1].amount).toBeNull();
-    expect(arr[0].type).toBe('Common Stock');
-    expect(arr[1].type).toBe('Debt Securities');
-    expect(result.rejectedFields.map((r) => r.path).sort()).toEqual([
-      'securityTypes[0].amount',
-      'securityTypes[1].amount',
-    ]);
-  });
-});
-
-describe('validateNumericGrounding — skip rules', () => {
-  it('skips when source doc is too short', () => {
-    const data: Record<string, unknown> = { offeringAmount: '$5B' };
-    const result = validateNumericGrounding(data, 'tiny', { formType: 'S-3' });
-    expect(result.rejectedFields).toEqual([]);
-    expect(data.offeringAmount).toBe('$5B');
-  });
-
-  it('skips emitted values shorter than 2 chars', () => {
-    const data: Record<string, unknown> = { offeringAmount: '$' };
-    const result = validateNumericGrounding(data, pad('blah'), { formType: 'S-3' });
-    expect(result.rejectedFields).toEqual([]);
-  });
-
-  it('does not walk prose fields (summary/whyItMatters/headline)', () => {
-    // Prose paths are NOT in WALK_PATHS, so any fabricated dollar inside
-    // them is left to ticker grounding (which handles prose contamination).
-    const data: Record<string, unknown> = {
-      summary: 'A fabricated $99B claim sits in this prose field',
-      whyItMatters: 'Another fabricated $77B never appears in the doc.',
-      headline: '$66B headline',
-    };
-    const result = validateNumericGrounding(
-      data,
-      pad('Real authorization $80B aggregate'),
-      { formType: 'S-3' },
-    );
-    expect(result.rejectedFields).toEqual([]);
-    expect(data.summary).toContain('$99B');
-  });
-
-  it('skips JSON-numeric typed values (only walks strings)', () => {
-    const data: Record<string, unknown> = {
-      offeringAmount: 99_999_999_999, // numeric, not string
-    };
-    const result = validateNumericGrounding(data, pad('Real authorization $80B'), { formType: 'S-3' });
-    expect(result.rejectedFields).toEqual([]);
-    expect(data.offeringAmount).toBe(99_999_999_999);
-  });
-
-  it('approximation tolerance: accepts $80B when source says "approximately $79.9 billion"', () => {
-    const data: Record<string, unknown> = { offeringAmount: '$80B' };
-    const result = validateNumericGrounding(
-      data,
-      pad('approximately $79.9 billion authorized for issuance'),
-      { formType: 'S-3' },
-    );
-    // Strict substring match might miss; approximation rule should still
-    // accept (or, falling back, accept because $79.9 isn't $80, but the
-    // intent is documented in the validator).
-    expect(data.offeringAmount).toBe('$80B');
-    // Relaxed assertion: validator may still reject if the exact "$80B"
-    // doesn't appear and the approximation rule is conservative. The key
-    // contract: this case should NOT cause a hard failure on unit-test
-    // CI — the validator is allowed to be conservative here.
-    void result;
-  });
-});
-
-describe('validateNumericGrounding — required-field unredactable guard', () => {
-  it('does not null Form 4 transactions[].shares when ungrounded; records unredactable', () => {
-    const data: Record<string, unknown> = {
-      transactions: [
-        { type: 'Sale', shares: '999,999', pricePerShare: '$50.00', code: 'S' },
-      ],
-    };
-    const sourceDoc = pad('Insider sold a small number of shares last week.');
-    const result = validateNumericGrounding(data, sourceDoc, { formType: '4' });
-    // shares is required, MUST NOT be nulled.
-    const tx = (data.transactions as Array<Record<string, unknown>>)[0];
-    expect(tx.shares).toBe('999,999');
-    expect(result.unredactablePaths).toContain('transactions[0].shares');
-  });
-});
-
-describe('validateNumericGrounding — chunked warn-only mode', () => {
-  it('counts violations but does NOT mutate data when chunked', () => {
-    const data: Record<string, unknown> = { offeringAmount: '$5B' };
-    const sourceDoc = pad('authorized $80B aggregate');
-    const result = validateNumericGrounding(data, sourceDoc, {
-      formType: 'S-3',
-      chunked: true,
-    });
-    expect(result.warnOnly).toBe(true);
-    expect(result.rejectedFields.length).toBe(1);
-    expect(data.offeringAmount).toBe('$5B'); // NOT mutated
-  });
-});
-
-describe('validateNumericGrounding — kill switch', () => {
-  it('short-circuits when ENRICHMENT_DISABLE_NUMERIC_GROUNDING=1', () => {
-    const original = process.env.ENRICHMENT_DISABLE_NUMERIC_GROUNDING;
-    process.env.ENRICHMENT_DISABLE_NUMERIC_GROUNDING = '1';
-    try {
-      const data: Record<string, unknown> = { offeringAmount: '$5B' };
-      const sourceDoc = pad('authorized $80B aggregate');
-      const result = validateNumericGrounding(data, sourceDoc, { formType: 'S-3' });
-      expect(result.rejectedFields).toEqual([]);
-      expect(data.offeringAmount).toBe('$5B');
-    } finally {
-      if (original === undefined) delete process.env.ENRICHMENT_DISABLE_NUMERIC_GROUNDING;
-      else process.env.ENRICHMENT_DISABLE_NUMERIC_GROUNDING = original;
-    }
-  });
-});
-
-describe('validateNumericGrounding — edge cases (P2 review additions)', () => {
-  it('comma-strip share count: 1.2M shares vs "1,234,567 shares"', () => {
-    const data: Record<string, unknown> = { sharesOffered: '1.2M shares' };
-    const sourceDoc = pad('the company sold 1,234,567 shares of common stock');
-    const result = validateNumericGrounding(data, sourceDoc, { formType: 'S-1' });
-    // 1.2M = 1,200,000. The doc has 1,234,567. These are not equal but the
-    // emitted value rounds. Approximation tolerance (5%): 1.2M vs 1.23M is
-    // within 3%, should accept.
-    expect(data.sharesOffered).toBe('1.2M shares');
-    void result;
-  });
-
-  it('rejects foreign currency: €500M emitted, no $500M in source', () => {
-    // The validator extracts via DOLLAR_PATTERN (which requires `$`), so
-    // €500M is treated as "no numeric content" and accepted. This is a
-    // documented limitation: euro-denominated values are not currently
-    // grounded. Captured here to surface it.
-    const data: Record<string, unknown> = { offeringAmount: '€500M' };
-    const sourceDoc = pad('issuance amount disclosed in EUR per attached schedule');
-    const result = validateNumericGrounding(data, sourceDoc, { formType: 'S-3' });
-    expect(data.offeringAmount).toBe('€500M');
-    void result;
-  });
-});
-
-describe('validateNumericGrounding — performance', () => {
-  it('reports durationMs on the result', () => {
-    const data: Record<string, unknown> = { offeringAmount: '$80B' };
-    const sourceDoc = pad('authorized $80,000,000,000 aggregate');
-    const result = validateNumericGrounding(data, sourceDoc, { formType: 'S-3' });
-    expect(typeof result.durationMs).toBe('number');
-    expect(result.durationMs).toBeGreaterThanOrEqual(0);
-  });
-
-  it('handles large source docs in under 250ms', () => {
-    const data: Record<string, unknown> = {
-      offeringAmount: '$80B',
-      shelfRegistration: { totalAuthorized: '$80B', remainingCapacity: '$60B' },
-      financialHighlights: Array.from({ length: 10 }, (_, i) => ({
-        metric: `M${i}`, value: `$${i + 1}B`, change: `+${i}%`,
-      })),
-    };
-    // 200KB source doc, mostly noise but containing the grounded values.
-    const noise = 'Lorem ipsum dolor sit amet '.repeat(8000);
-    const sourceDoc = `${noise}\nauthorized $80,000,000,000 aggregate; remaining $60B; ${
-      Array.from({ length: 10 }, (_, i) => `metric ${i}: $${i + 1}B at +${i}%`).join('. ')
-    }\n${noise}`;
-    const t0 = Date.now();
-    const result = validateNumericGrounding(data, sourceDoc, { formType: '10-Q' });
-    const elapsed = Date.now() - t0;
-    expect(elapsed).toBeLessThan(250);
-    expect(result.durationMs).toBeLessThan(250);
-  });
-});
